#31 [refactor] backend: extract CarRepository from domain services to infrastructure layer
Move mock data and repository logic out of domain/services into infrastructure/car.repository.js,
add CarRepository interface to domain lib, and wire composition in endpoints/index.js.
