+++ apps/backend/src/endpoints/index.js
@@ -1,7 +1,13 @@
-import domain from '../domain/services';
+import { createCarService } from '../domain/services';
+import { createMockCarRepository } from '../infrastructure/car.repository';
 
 import { createCarEndpoint } from './car';
 
+const carRepository = createMockCarRepository();
+const domain = {
+  cars: createCarService({ cars: carRepository }),
+};
+
 export default {
   cars: createCarEndpoint(domain),
-}; 
+};
