+++ libs/domain/src/index.ts
@@ -1,2 +1,3 @@
 export * from "./entities";
 export * from "./value-object";
+export * from "./repositories";
