+++ libs/domain/src/repositories/index.ts
@@ -1,0 +1,3 @@
+export * from './crud';
+export * from './car.repository';
+export * from './page.repository';
